requirements.txt
requests~=2.24.0
pymongo==3.0.1
bottle==0.12.18
pytest==5.4.3
argcomplete==1.12.0
pyhamcrest==2.0.2
prettytable==0.7.2
